requirements.txt
python-dotenv==1.0.0
requests==2.28.2
Flask==2.2.3
pymongo==4.3.3
wand==0.6.11
PyPDF2-3.0.1